* Return the filters_used.
