+++ python/agActor/focus.py
@@ -5,7 +5,7 @@ from numpy._typing import ArrayLike
 
 from kawanomoto import FieldAcquisitionAndFocusing
 from agActor.opdb import opDB as opdb
-from agActor.field_acquisition import semi_axes
+from agActor.utils import semi_axes
 from agActor.utils import _KEYMAP, filter_kwargs, map_kwargs
